fix(session): viewtransition + body attributes
